Extracted component and implemented keyhandling logic
